help.py(class Help) ->info.py(class Info)
